DeskBooking is only used here, this is tidier
